refactor(api): move machine controller to separate crate
