+++ crates/api/src/state_controller/mod.rs
@@ -16,4 +16,3 @@
  */
 
 pub mod common_services;
-pub mod machine;
