+++ crates/api/src/tests/dpf/reprovisioning.rs
@@ -26,13 +26,13 @@ use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
 use carbide_dpf::DpuPhase;
+use carbide_machine_controller::dpf::{DpfOperations, MockDpfOperations};
 use carbide_uuid::machine::MachineId;
 use model::machine::{
     DpfState, DpuReprovisionStates, InstanceState, ManagedHostState, ReprovisionState,
 };
 use tokio::time::timeout;
 
-use crate::state_controller::machine::dpf::{DpfOperations, MockDpfOperations};
 use crate::tests::common::api_fixtures::{
     TestEnvOverrides, TestManagedHost, create_managed_host_with_dpf,
     create_managed_host_with_dpf_multi, create_test_env_with_overrides, get_config,
