+++ crates/api/src/tests/host_bmc_firmware_test.rs
@@ -22,6 +22,8 @@ use std::os::unix::fs::PermissionsExt;
 use std::str::FromStr;
 use std::time::Duration;
 
+use carbide_machine_controller::config::{FirmwareGlobal, TimePeriod};
+use carbide_machine_controller::handler::MAX_FIRMWARE_UPGRADE_RETRIES;
 use carbide_preingestion_manager::PreingestionManager;
 use carbide_redfish::libredfish::test_support::RedfishSimAction;
 use carbide_uuid::machine::MachineId;
@@ -49,10 +51,8 @@ use tokio::time::sleep;
 use tonic::Request;
 
 use crate::CarbideResult;
-use crate::cfg::file::{CarbideConfig, TimePeriod};
+use crate::cfg::file::CarbideConfig;
 use crate::machine_update_manager::MachineUpdateManager;
-use crate::state_controller::machine::config::FirmwareGlobal;
-use crate::state_controller::machine::handler::MAX_FIRMWARE_UPGRADE_RETRIES;
 use crate::tests::common;
 use crate::tests::common::api_fixtures::managed_host::HardwareInfoTemplate;
 use crate::tests::common::api_fixtures::{
