+++ crates/api/src/tests/machine_creator.rs
@@ -20,6 +20,7 @@ use std::net::IpAddr;
 use std::str::FromStr;
 use std::sync::Arc;
 
+use carbide_machine_controller::handler::MachineStateHandlerBuilder;
 use carbide_site_explorer::MachineCreator;
 use carbide_site_explorer::config::SiteExplorerConfig;
 use carbide_site_explorer::errors::SiteExplorerError;
@@ -39,7 +40,6 @@ use rpc::{BlockDevice, DiscoveryData, DiscoveryInfo, MachineDiscoveryInfo};
 use tonic::Request;
 
 use crate::cfg::file::DpuConfig as InitialDpuConfig;
-use crate::state_controller::machine::handler::MachineStateHandlerBuilder;
 use crate::tests::common;
 use crate::tests::common::api_fixtures::TestEnvOverrides;
 use crate::tests::common::api_fixtures::dpu::DpuConfig;
