+++ crates/api/src/tests/machine_update_manager.rs
@@ -20,6 +20,7 @@ use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
 use async_trait::async_trait;
+use carbide_machine_controller::health_report::create_host_update_health_report;
 use carbide_uuid::machine::MachineId;
 use common::api_fixtures::create_test_env;
 use figment::Figment;
@@ -37,7 +38,6 @@ use crate::CarbideResult;
 use crate::cfg::file::CarbideConfig;
 use crate::machine_update_manager::MachineUpdateManager;
 use crate::machine_update_manager::machine_update_module::MachineUpdateModule;
-use crate::state_controller::machine::health_report::create_host_update_health_report;
 use crate::tests::common;
 use crate::tests::common::api_fixtures::create_managed_host;
