+++ crates/api/src/tests/mod.rs
@@ -75,6 +75,7 @@ mod machine_interfaces;
 mod machine_metadata;
 mod machine_network;
 mod machine_power;
+mod machine_setup;
 mod machine_states;
 mod machine_topology;
 pub mod machine_update_manager;
