+++ crates/machine-controller/Cargo.toml
@@ -1,0 +1,76 @@
+#
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+[package]
+name = "carbide-machine-controller"
+version = "0.0.0"
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+
+[features]
+default = []
+test-support = ["dep:mockall"]
+
+[dependencies]
+bmc-vendor = { path = "../bmc-vendor" }
+carbide-api-db = { path = "../api-db", default-features = false }
+carbide-api-model = { path = "../api-model", default-features = false }
+carbide-dpf = { path = "../dpf", default-features = false }
+carbide-health-report = { path = "../health-report", default-features = false }
+carbide-health-metrics = { path = "../health-metrics" }
+carbide-utils = { path = "../utils", default-features = false }
+carbide-firmware = { path = "../firmware", default-features = false }
+carbide-ipmi = { path = "../ipmi", default-features = false }
+carbide-measured-boot = { path = "../measured-boot", default-features = false }
+carbide-redfish = { path = "../redfish", default-features = false }
+# TODO: RPC is only used to serialize / deserialize of
+# ScoutFirmwareUpgradeTask. If we can find a way to do it differently
+# we can improve build performance significantly.
+carbide-rpc = { path = "../rpc", default-features = false }
+carbide-secrets = { path = "../secrets" }
+carbide-state-controller-common = { path = "../state-controller-common", default-features = false }
+carbide-uuid = { path = "../uuid", default-features = false }
+config-version = { path = "../config-version", default-features = false }
+state-controller = { path = "../state-controller" }
+
+async-trait = { workspace = true }
+chrono = { workspace = true }
+duration-str = { workspace = true }
+eyre = { workspace = true }
+futures = { workspace = true }
+futures-util = { workspace = true }
+itertools = { workspace = true }
+lazy_static = { workspace = true }
+libredfish = { workspace = true }
+mac_address = { workspace = true }
+mockall = { workspace = true, optional = true }
+opentelemetry = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+sqlx = { workspace = true }
+tracing = { workspace = true }
+tokio = { workspace = true }
+uuid = { workspace = true, features = ["v4", "serde"] }
+version-compare = { workspace = true }
+
+[dev-dependencies]
+figment = { workspace = true, features = ["env", "test", "toml"] }
+regex = { workspace = true }
+lazy_static = { workspace = true }
+
+[lints]
+workspace = true
