+++ crates/machine-controller/src/config/bom_validation.rs
