+++ crates/machine-controller/src/config/power_manager.rs
