+++ crates/machine-controller/src/handler/helpers.rs
