+++ crates/machine-controller/src/handler/power.rs
@@ -25,10 +25,8 @@ use model::power_manager::{
 };
 use state_controller::state_handler::{StateHandlerContext, StateHandlerError};
 
-use crate::state_controller::machine::context::MachineStateHandlerContextObjects;
-use crate::state_controller::machine::handler::{
-    PowerOptionConfig, handler_host_power_control, host_power_state,
-};
+use crate::context::MachineStateHandlerContextObjects;
+use crate::handler::{PowerOptionConfig, handler_host_power_control, host_power_state};
 
 // If power state is Paused and Reset, state machine can't take any decision on it.
 // Ignore power manager with a log and moved to state machine.
