+++ crates/machine-controller/src/health_report.rs
