+++ crates/machine-controller/src/write_ops.rs
